Refactoring: TUF* BaseModel in a commom package
This commit makes the TUF BaseModels reusable in common package.

This models will be used to implement `/api/v1/metadata`
