+++ repository_service_tuf_api/bootstrap.py
@@ -4,8 +4,7 @@
 
 import json
 import logging
-from enum import Enum
-from typing import Dict, List, Literal, Optional
+from typing import Dict, Literal, Optional
 
 from fastapi import HTTPException, status
 from pydantic import BaseModel, Field
@@ -16,88 +15,12 @@ from repository_service_tuf_api import (
     repository_metadata,
     settings_repository,
 )
+from repository_service_tuf_api.common_models import (
+    BaseErrorResponse,
+    Roles,
+    TUFMetadata,
+)
 from repository_service_tuf_api.config import save_settings
-
-
-class Roles(Enum):
-    ROOT = "root"
-    TARGETS = "targets"
-    SNAPSHOT = "snapshot"
-    TIMESTAMP = "timestamp"
-    BINS = "bins"
-
-    @classmethod
-    def values(cls) -> List[str]:
-        return Literal["root", "targets", "snapshot", "timestamp", "bins"]
-
-
-class BaseErrorResponse(BaseModel):
-    error: str = Field(description="Error message")
-    details: Optional[Dict[str, str]] = Field(description="Error details")
-    code: Optional[int] = Field(description="Error code if available")
-
-
-class TUFSignedDelegationsRoles(BaseModel):
-    name: str
-    terminating: bool
-    keyids: List[str]
-    threshold: int
-    paths: Optional[List[str]]
-    path_hash_prefixes: Optional[List[str]]
-
-
-class TUFSignedDelegationsSuccinctRoles(BaseModel):
-    bit_length: int = Field(gt=0, lt=15)
-    name_prefix: str
-    keyids: List[str]
-    threshold: int
-
-
-class TUFKeys(BaseModel):
-    keytype: str
-    scheme: str
-    keyval: Dict[Literal["public", "private"], str]
-
-
-class TUFSignedDelegations(BaseModel):
-    keys: Dict[str, TUFKeys]
-    roles: Optional[List[TUFSignedDelegationsRoles]]
-    succinct_roles: Optional[TUFSignedDelegationsSuccinctRoles]
-
-
-class TUFSignedMetaFile(BaseModel):
-    version: int
-
-
-class TUFSignedRoles(BaseModel):
-    keyids: List[str]
-    threshold: int
-
-
-class TUFSigned(BaseModel):
-    type: str
-    version: int
-    spec_version: str
-    expires: str
-    keys: Optional[Dict[str, TUFKeys]]
-    consistent_snapshot: Optional[bool]
-    roles: Optional[Dict[Roles.values(), TUFSignedRoles]]
-    meta: Optional[Dict[str, TUFSignedMetaFile]]
-    targets: Optional[Dict[str, str]]
-    delegations: Optional[TUFSignedDelegations]
-
-    class Config:
-        fields = {"type": "_type"}
-
-
-class TUFSignatures(BaseModel):
-    keyid: str
-    sig: str
-
-
-class TUFMetadata(BaseModel):
-    signatures: List[TUFSignatures]
-    signed: TUFSigned
 
 
 class ServiceSettings(BaseModel):
